test: fix failing tests
